Can not create a Path from an empty string
Caused by: org.apache.hive.hcatalog.common.HCatException : 2006 : Error adding partition to metastore. Cause : java.lang.IllegalArgumentException: Can not create a Path from an empty string
	at org.apache.hive.hcatalog.mapreduce.FileOutputCommitterContainer.registerPartitions(FileOutputCommitterContainer.java:975)
	at org.apache.hive.hcatalog.mapreduce.FileOutputCommitterContainer.commitJob(FileOutputCommitterContainer.java:250)
	... 20 more

+++ hcatalog/core/src/main/java/org/apache/hive/hcatalog/mapreduce/FileOutputCommitterContainer.java
@@ -891,7 +891,9 @@ class FileOutputCommitterContainer extends OutputCommitterContainer {
               if (src.toString().matches(".*" + Path.SEPARATOR + SCRATCH_DIR_NAME + "\\d\\.?\\d+.*")){
                 // src is scratch directory, need to trim the part key value pairs from path
                 String diff = StringUtils.difference(src.toString(), dest.toString());
-                fs.delete(new Path(StringUtils.substringBefore(src.toString(), diff)), true);
+                if (StringUtils.isNotBlank(diff)) {
+                  fs.delete(new Path(StringUtils.substringBefore(src.toString(), diff)), true);
+                }
               } else {
                 fs.delete(src, true);
               }
